...G...... [ZBXNEXT-1287] removed numeric counter path conversion, fixed warnings

+++ src/zabbix_agent/cpustat.c
@@ -136,7 +136,7 @@ int	init_cpu_collector(ZBX_CPUS_STAT_DATA *pcpus)
 	cpe.szObjectName = get_counter_name(PCI_PROCESSOR);
 	cpe.szInstanceName = cpu;
 	cpe.szParentInstance = NULL;
-	cpe.dwInstanceIndex = -1;
+	cpe.dwInstanceIndex = (DWORD)-1;
 	cpe.szCounterName = get_counter_name(PCI_PROCESSOR_TIME);
 
 	for (idx = 0; idx <= pcpus->count; idx++)
